refactor(dependencies): 收敛 Windows 依赖适配逻辑
