+++ pinvou3-app/src-tauri/src/features/files/file_ingest.rs
@@ -116,13 +116,11 @@ pub fn check_dependencies() -> Vec<DependencyCheckItem> {
             crate::platform::os::pandoc_dependency_packages(),
         ));
     }
-    if crate::features::dependencies::dependency_check_policy().include_voice_runtime {
-        items.push(item(
-            "voice_asr",
-            crate::features::voice::asr_tool_exists(),
-            crate::features::voice::asr_dependency_packages(),
-        ));
-    }
+    items.push(item(
+        "voice_asr",
+        crate::features::voice::asr_tool_exists(),
+        crate::features::voice::asr_dependency_packages(),
+    ));
     items.push(item("office_legacy", libreoffice, "libreoffice"));
     if crate::platform::os::show_ocr_dependency_check() {
         items.push(item(
@@ -2738,7 +2736,6 @@ mod tests {
         let has_pandoc = deps.iter().any(|item| item.key == "office_modern");
         let has_ocr = deps.iter().any(|item| item.key == "ocr");
         let has_archive = deps.iter().any(|item| item.key == "archive");
-        let has_voice_asr = deps.iter().any(|item| item.key == "voice_asr");
         assert_eq!(has_pdf, crate::platform::os::show_pdf_dependency_check());
         assert_eq!(
             has_pandoc,
@@ -2748,10 +2745,6 @@ mod tests {
         assert_eq!(
             has_archive,
             crate::platform::os::show_archive_dependency_check()
-        );
-        assert_eq!(
-            has_voice_asr,
-            crate::features::dependencies::dependency_check_policy().include_voice_runtime
         );
 
         if !crate::platform::os::show_pdf_dependency_check() {
